коррекция типа выгрузки сикурити левела

+++ src/server/game/World/World.cpp
@@ -2984,7 +2984,7 @@ void World::LoadDBAllowedSecurityLevel()
     PreparedQueryResult result = LoginDatabase.Query(stmt);
 
     if (result)
-        SetPlayerSecurityLimit(AccountTypes(result->Fetch()->GetUInt8()));
+        SetPlayerSecurityLimit(AccountTypes(result->Fetch()->GetFloat()));
 }
 
 void World::SetPlayerSecurityLimit(AccountTypes _sec)
